Initial version added
this is a crude staff and crude login system just for a prototype. this sis to test the clock in and out saving algorythm and staff adding to the program. this is by no means the finished thing. i still need to add many things and iron out more bigs and add validation and proper verification